easyEval2/requirements.txt
# 核心依赖
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0

# 数据处理
json5>=0.9.0
pydantic>=2.0.0

# 日志和工具
rich>=13.0.0
click>=8.0.0
tqdm>=4.64.0

# 开发和测试
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# 可选依赖（用于高级功能）
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.6.0
seaborn>=0.12.0